justright-server/requirements.txt
Django==3.2.4
djangorestframework==3.13.1
drf_yasg==1.20.0
PyMySQL==1.0.2
requests==2.25.0
markdown
django-filter
django-cors-headers
PyJWT==1.7.1
PyEmail
email-to
openpyxl==3.0.10
pandas==1.3.5